Validar si existen tareas antes de eliminar un proyecto

+++ src/controllers/projects.controller.js
@@ -46,6 +46,14 @@ export const getProject = async (req, res) => {
 
 export const deleteProject = async (req, res) => {
     try {
+
+        let [tareas_dependientes] = await pooldb.query("SELECT FROM tbl_tarea WHERE id_proyecto = ? ", [req.params.id]);
+        if(rows.affectedRows > 0){
+            return res.status(400).json({
+                message: "El proyecto aun tiene tareas sin terminar"
+            })
+        }
+
         let [rows] = await pooldb.query("DELETE FROM tbl_proyecto WHERE id = ?", [req.params.id]);
         if (rows.affectedRows <= 0)
             return res.status(404).json({
